+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 *Latest on top*
 
+* ## [Modded tags &raquo; each](anton.cms@ya.ru__tags-modded/each)
+   Original tag expanded with 4 more parameters (***reverse, count_only, skip_empty, limit***).
 * ## [New tags &raquo; delete_file](anton.cms@ya.ru__tags-new/delete_file)
    Plus can delete PHP's error_log.
 * ## [**Addon "Func-on-demand"**](https://github.com/trendoman/Tweakus-Dilectus/tree/main/anton.cms@ya.ru__func-on-demand)
-   Autoloading and caching of 'func' snippets.
+   **Autoloading** and **caching** of 'func' snippets.
 * ## [New tags &raquo; unset](anton.cms@ya.ru__tags-new/unset)
    Remove a variable or array key.
 * ## [New tags &raquo; admin_create_link](anton.cms@ya.ru__tags-new/admin_create_link)